+++ src/callbacks/perturb_adam.cpp
@@ -26,7 +26,7 @@
 
 #include "lbann/callbacks/perturb_adam.hpp"
 #include "lbann/proto/proto_common.hpp"
-#include "lbann/utils/random.hpp"
+#include "lbann/utils/random_number_generators.hpp"
 
 #include <callbacks.pb.h>
